feat(dapp): fix parentnftid in hook call
