+++ dapp/src/lib/auction/utils.ts
@@ -4,11 +4,7 @@
 import { bcs } from '@iota/iota-sdk/bcs';
 import { blake2b } from '@noble/hashes/blake2';
 import { bytesToHex } from '@noble/hashes/utils';
-import { useMemo } from 'react';
 
-import { NameRecordData, useNameRecord } from '@/hooks';
-
-import { getNamePermissions } from '../utils/names';
 import { DomainBcs } from './bcs';
 
 /**
@@ -55,71 +51,4 @@ export function deriveAuctionDynamicFieldId(
 export function createDomainFromName(domainName: string): { labels: string[] } {
     const labels = domainName.split('.');
     return { labels: labels.reverse() };
-}
-
-function getSubdomainLevel(name: string): number {
-    const parts = name.split('.');
-    return Math.max(0, parts.length - 2);
-}
-
-function getDirectParent(name: string): string | null {
-    const parts = name.split('.');
-    if (parts.length <= 2) {
-        return null;
-    }
-    return parts.slice(1).join('.');
-}
-/**
- * Helper function to validate subdomain permissions based on hierarchical inheritance rules.
- *
- * - Root domains (domain.iota): Cannot configure permissions (always implicitly enabled)
- * - First-level subdomains (sub.domain.iota): Can configure permissions freely
- * - Deep subdomains (test.sub.domain.iota): Must respect parent's permission restrictions
- *
- * For deep subdomains, if a parent has a permission disabled (e.g., allowTimeExtension: false),
- * the child subdomain cannot enable that permission, ensuring hierarchical permission inheritance.
- *
- * @param fullSubdomainName - The complete subdomain name to validate (e.g., "test.sub.domain.iota")
- * @returns Parent permissions: canModifyTimeExtension and canModifyChildCreation
- */
-export function useSubdomainPermissionsValidation(fullSubdomainName: string) {
-    const subdomainLevel = getSubdomainLevel(fullSubdomainName);
-    const directParent = getDirectParent(fullSubdomainName);
-
-    const { data: parentNameRecord } = useNameRecord(directParent || '');
-
-    const { data } =
-        subdomainLevel >= 2 ? useNameRecord(directParent!) : useNameRecord(fullSubdomainName);
-    const nameRecord = data as Extract<NameRecordData, { type: 'unavailable' }> | undefined;
-
-    const parentPermissions =
-        parentNameRecord && nameRecord?.nameRecord
-            ? getNamePermissions(nameRecord.nameRecord)
-            : undefined;
-
-    const canModifyTimeExtension = useMemo(() => {
-        if (subdomainLevel === 0) return false;
-        if (subdomainLevel === 1) return true;
-        if (subdomainLevel >= 2) {
-            if (!parentPermissions) return false;
-            return parentPermissions.allowTimeExtension === true;
-        }
-
-        return false;
-    }, [fullSubdomainName, subdomainLevel, directParent, parentPermissions]);
-
-    const canModifyChildCreation = useMemo(() => {
-        if (subdomainLevel === 0) return false;
-        if (subdomainLevel === 1) return true;
-        if (subdomainLevel >= 2) {
-            if (!parentPermissions) return false;
-            return parentPermissions?.allowChildCreation === true;
-        }
-        return false;
-    }, [subdomainLevel, parentPermissions]);
-
-    return {
-        canModifyTimeExtension,
-        canModifyChildCreation,
-    };
 }
